backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
pydantic==2.9.0
pydantic-settings==2.5.0
pydantic[email]==2.9.0
python-multipart==0.0.6
websockets==12.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1
python-dotenv==1.0.0
httpx==0.27.2
SpeechRecognition==3.14.5
pydub==0.25.1
pillow==10.1.0
numpy==1.26.2
motor==3.3.2
pymongo==4.6.3
cryptography>=41.0.0
slowapi==0.1.9
